feat(services/obs): support sink (#2578)

+++ core/src/services/obs/core.rs
@@ -147,7 +147,7 @@ impl ObsCore {
     pub fn obs_put_object_request(
         &self,
         path: &str,
-        size: Option<usize>,
+        size: Option<u64>,
         content_type: Option<&str>,
         cache_control: Option<&str>,
         body: AsyncBody,
